Task123
Menu works on mac as expected; not apart of application window, but at the top of the screen. I think that’s fine. I did add some menus I saw were there, but not connected. Interestingly, on macOS there are some default things added the editmenu, but are not added to the same menu under File.